UI: Move tableSort func to utils
