+++ ui/src/utils/index.js
@@ -2,6 +2,7 @@ import filterMethod from './filterMethod';
 import stringToInitials from './stringToInitials';
 import stringToColor from './stringToColor';
 import sortMethod from './sortMethod';
+import tableOrder from './tableOrder';
 import uuidv4 from './uuidv4';
 
 export {
@@ -9,5 +10,6 @@ export {
     sortMethod,
     stringToInitials,
     stringToColor,
+    tableOrder,
     uuidv4,
 }
